Initial push for codesharing project
This is the initial push for the codesharing project. All the work was initially done locally, then moved to this single folder. It is possible that errors may occur after moving data. If that is the case, a second push should be seen shortly.

Code is currently very messy, with errors commented out. This should be fixed in the next push, as well as the poor graph.