Prettier code now
I lied in the comments of my last code push, the issue in the naming of display_field was actually fixed in that push. The README has been updated to reflect this change. All code with errors has been commented out, and everything should now be clean and functional.

This (hopefully) is the final push of this project unless more instructions come out.